added triple shot
made the triple shot power up work with lasers and have the same time duration as lasers (although they are separate variables and are treated separately so this can be changed easily).
Also made the sounds different (higher pitched for more shots fired during triple shot). This change is completely subjective and can be changed back simply if we want.