Setup for future work
I removed all logic regarding the bounds of the window. Instead, the "bounds of the window" are simply walls places on the boundary of the window. 

The next "big thing" we should work on is getting a "whole level" that is larger than the window and have the window follow the player. This will actually be close to a "real game" rather than a fun little box for mechanic testing